pyproject.toml
[project]
authors = [{name = "cafe small house", email = "ybaram40@gmail.com"}]
dependencies = [
  "django>=5.2,<6",
  "django-ninja>=1.4.0",
  "environs>=11.0.0,<12",
  "pillow>=10.4.0,<11",
  "psycopg2>=2.9.9,<3",
]
description = ""
license = "MIT"
name = "airbnb"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"

[dependency-groups]
dev = [
  "djlint>=1.36.4",
  "ruff>=0.9.0",
]

[tool.djlint]
format_css = true
format_js = true
indent = 4
max_line_length = 160
profile = "django"

[tool.ruff]
line-length = 160
indent-width = 4

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = ["E", "F", "I"]
